format files per clang

+++ include/processor.h
@@ -7,17 +7,17 @@ class Processor {
 
   // TODO: Declare any necessary private members
  private:
- // CPU calc aux
-  int last_user = 0;
-  int last_nice = 0;
-  int last_system = 0;
-  int last_idle = 0;
-  int last_iowait = 0;
-  int last_irq = 0;
-  int last_softirq = 0;
-  int last_steal = 0;
-  int last_guest = 0;
-  int last_gest_nice = 0;
+  // CPU calc aux
+  long int last_user = 0;
+  long int last_nice = 0;
+  long int last_system = 0;
+  long int last_idle = 0;
+  long int last_iowait = 0;
+  long int last_irq = 0;
+  long int last_softirq = 0;
+  long int last_steal = 0;
+  long int last_guest = 0;
+  long int last_gest_nice = 0;
 };
 
 #endif
